Create binary_gap1.py
In a positive integer N, a binary gap is defined as the longest sequence of continuous zeros that is flanked by ones on both ends in N's binary format.

Take, for instance, the integer 9, which in binary is expressed as 1001, and it showcases a binary gap of length 2. The integer 529, with a binary notation of 1000010001, features two binary gaps: one with a length of 4 and another with a length of 3. The integer 20, represented in binary as 10100, includes a single binary gap of length 1. Conversely, the integer 15, which is represented as 1111 in binary, does not possess any binary gaps. Similarly, the integer 32, with a binary representation of 100000, lacks binary gaps.