Create PermMissingElem.py
A non-empty array A consisting of N integers is given. Array A represents numbers on a tape.
Any integer P, such that 0 < P < N, splits this tape into two non-empty parts
In other words, it is the absolute difference between the sum of the first part and the sum of the second part.
Write a function that, given a non-empty array A of N integers, returns the minimal difference that can be achieved.