Create matrix_decoding.py
Takes in matrix in form given below.
and decodes the matrix in the form of a string, considering only alpha-numeric characters. the program does not ignore other characters if they are not prefixed or suffixed by an alpha-numeric. 
"do not use 'if' statement" is part of challenge. 
input:
7 3
Tsi
h%x
i #
sM 
$a 
#t%
ir!
 output:

This is Matrix#  %!